First cleanup of code

+++ app/src/main/java/com/example/roleplaybanking/database/DatabaseCon.java
@@ -4,7 +4,6 @@ package com.example.roleplaybanking.database;
 import android.content.SharedPreferences;
 import android.util.Log;
 
-import com.example.roleplaybanking.R;
 import com.example.roleplaybanking.controllers.AccountSelectionActivity;
 import com.example.roleplaybanking.controllers.CreateNewActivity;
 import com.example.roleplaybanking.structures.Account;
@@ -35,7 +34,6 @@ public class DatabaseCon {
 
     private NutzerClass user;
     private String PW;
-    private boolean success = false;
     public final ArrayList<Account> accounts = new ArrayList<>();
     private final ArrayList<Transaction> transactions = new ArrayList<>();
     private final ArrayList<Recipients> recipients = new ArrayList<>();
@@ -55,19 +53,18 @@ public class DatabaseCon {
     }
 
     public void registerUser(String Name, Number Id, String User, String UserPW) {
-        NutzerClass newuser = new NutzerClass(Name, Id, User, UserPW);
+        NutzerClass newuser = new NutzerClass(Name, Id, UserPW);
         database.collection("Nutzer").document(User).set(newuser);
         DocMenge.update("nutzer", (long) Id + 1);
         user = newuser;
-        success = true;
     }
 
     public void registerAccount(CreateNewActivity act, String Game, Number Geld, String Kontonamen, boolean isnewGame, double dbalance) {
         DocMenge.get().addOnSuccessListener(documentSnapshot -> {
             Map<String, Object> m = documentSnapshot.getData();
-            if(isnewGame){
-                registerGame(Game, (Number)m.get("kontos"), dbalance);
-            }
+            if (isnewGame)
+                registerGame(Game, (Number) m.get("kontos"), dbalance);
+
             registerAccount(Game, Geld, (long) m.get("kontos"), Kontonamen, user.getNutzerID());
             if (act != null)
                 act.closeActivityWhenDone();
@@ -75,19 +72,20 @@ public class DatabaseCon {
     }
 
     public void registerAccount(String Game, Number Geld, Number KontoID, String Kontonamen, Number NutzerID) {
-        Account newAcc = new Account(Game, Kontonamen,  Double.parseDouble(Geld.toString()), KontoID);
+        Account newAcc = new Account(Game, Kontonamen, Double.parseDouble(Geld.toString()), KontoID);
         Map<String, Object> m = new HashMap<>();
         m.put("Game", Game);
         m.put("Geld", Geld);
         m.put("KontoID", KontoID);
         m.put("Kontoname", Kontonamen);
         m.put("Nutzer", NutzerID);
+
         database.collection("Konten").document(KontoID.toString()).set(m);
         DocMenge.update("kontos", (long) KontoID + 1);
         accounts.add(newAcc);
+
         Recipients R = new Recipients(Kontonamen, Game, KontoID);
         recipients.add(R);
-        success = true;
     }
 
     public void registerTransaction(Number Betrag, Number Empfaenger, String Notiz, Number Nutzerkonto, Timestamp time, Account fromAcc) {
@@ -99,16 +97,14 @@ public class DatabaseCon {
 
     public void registerTransaction(Number Betrag, Number Empfaenger, Number HistoryID, String Notiz, Number Nutzer, Number Nutzerkonto, Timestamp time, Account fromAcc) {
         int e;
-        for(e=0; e < recipients.size(); e++){
-            if(recipients.get(e).KontoID.equals(Empfaenger)){
+        for (e = 0; e < recipients.size(); e++) {
+            if (recipients.get(e).KontoID.equals(Empfaenger))
                 break;
-            }
         }
         int n;
-        for(n=0; n < recipients.size(); n++){
-            if(recipients.get(n).KontoID.equals(Nutzerkonto)){
+        for (n = 0; n < recipients.size(); n++) {
+            if (recipients.get(n).KontoID.equals(Nutzerkonto))
                 break;
-            }
         }
         Transaction newTran = new Transaction(recipients.get(n).Name, recipients.get(e).Name, (double) Betrag, time, Notiz, HistoryID, fromAcc.gameName);
         Map<String, Object> m = new HashMap<>();
@@ -119,11 +115,12 @@ public class DatabaseCon {
         m.put("Nutzer", Nutzer);
         m.put("Nutzerkonto", Nutzerkonto);
         m.put("SendeZeit", time);
+
         database.collection("History").document(HistoryID.toString()).set(m);
         DocMenge.update("historys", (long) HistoryID + 1);
+
         transactions.add(newTran);
         fromAcc.AccountHistory.add(newTran);
-        success = true;
     }
 
     public void registerGame(String name, Number Admin, double dbalance) {
@@ -139,37 +136,34 @@ public class DatabaseCon {
         m.put("GameID", GameID);
         m.put("Name", name);
         m.put("DefaultBalance", dbalance);
+
         database.collection("Game").document(GameID.toString()).set(m);
         DocMenge.update("games", (long) GameID + 1);
         Game a = new Game(name, Admin, dbalance);
         games.add(a);
     }
 
-    public void connect(AccountSelectionActivity activity, String Name, Number Id, String User, String UserPW) {
-        user = new NutzerClass(Name, Id, User, UserPW);
-        if (!(user.getNutzerPW().equals(PW))) {
+    public void connect(AccountSelectionActivity activity, String Name, Number Id, String UserPW) {
+        user = new NutzerClass(Name, Id, UserPW);
+        if (!(user.getNutzerPW().equals(PW)))
             user = null;
-            success = false;
-        } else {
-            success = true;
-        }
-        System.out.println("Start");
+
         this.connectGames(activity);
     }
 
     public void connectUser(AccountSelectionActivity activity, SharedPreferences sharedPreferences) {
         this.sharedPreferences = sharedPreferences;
-        String currentUserID = getFromDeviceStorage("UserID", "0123456789");
-        String currentUserPW = getFromDeviceStorage("UserPW", "0123456789");
+        String currentUserID = getFromDeviceStorage("UserID");
+        String currentUserPW = getFromDeviceStorage("UserPW");
 
         DocumentReference DocUser = ColUser.document(currentUserID);
         PW = currentUserPW;
         DocUser.get().addOnSuccessListener(documentSnapshot -> {
             Map<String, Object> m = documentSnapshot.getData();
-            if (m == null) {
+            if (m == null)
                 registerUser(currentUserID, currentUserID, currentUserPW);
-            } else
-                connect(activity, m.get("name").toString(), (long) m.get("nutzerID"), m.get("nutzerName").toString(), m.get("nutzerPW").toString());
+            else
+                connect(activity, m.get("name").toString(), (long) m.get("nutzerID"), m.get("nutzerPW").toString());
         });
     }
 
@@ -177,7 +171,6 @@ public class DatabaseCon {
         if (user == null)
             return;
 
-        System.out.println("Start Accounts");
         ColAccount.whereEqualTo("Nutzer", user.getNutzerID()).get().addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
             final AccountSelectionActivity act = activity;
 
@@ -187,9 +180,7 @@ public class DatabaseCon {
                     for (QueryDocumentSnapshot document : task.getResult()) {
                         Map<String, Object> m = document.getData();
                         addAccount(m.get("Game").toString(), m.get("Kontoname").toString(), Double.parseDouble(m.get("Geld").toString()), (Number) m.get("KontoID"));
-                        System.out.print(m.get("KontoID").toString());
-                        System.out.println("Start");
-                        connectTransactions((Number)m.get("KontoID"), m.get("Game").toString());
+                        connectTransactions((Number) m.get("KontoID"), m.get("Game").toString());
                     }
                     if (act != null)
                         act.notifyDBConnectionEstablished();
@@ -208,7 +199,7 @@ public class DatabaseCon {
             if (task.isSuccessful()) {
                 for (QueryDocumentSnapshot document : task.getResult()) {
                     Map<String, Object> m = document.getData();
-                    addTransaction((Number)m.get("Nutzerkonto"), (Number)m.get("Empfaenger"),  Double.parseDouble(m.get("Betrag").toString()), (Timestamp) m.get("SendeZeit"), m.get("Notiz").toString(), (Number)m.get("HistoryID"), Gamename);
+                    addTransaction((Number) m.get("Nutzerkonto"), (Number) m.get("Empfaenger"), Double.parseDouble(m.get("Betrag").toString()), (Timestamp) m.get("SendeZeit"), m.get("Notiz").toString(), (Number) m.get("HistoryID"), Gamename);
                 }
             }
         });
@@ -217,7 +208,7 @@ public class DatabaseCon {
                 for (QueryDocumentSnapshot document : task.getResult()) {
                     Log.d("ConnectTrans", document.getData().toString());
                     Map<String, Object> m = document.getData();
-                    addTransaction((Number)m.get("Nutzerkonto"), (Number)m.get("Empfaenger"), Double.parseDouble(m.get("Betrag").toString()), (Timestamp) m.get("SendeZeit"), m.get("Notiz").toString(), (Number)m.get("HistoryID"), Gamename);
+                    addTransaction((Number) m.get("Nutzerkonto"), (Number) m.get("Empfaenger"), Double.parseDouble(m.get("Betrag").toString()), (Timestamp) m.get("SendeZeit"), m.get("Notiz").toString(), (Number) m.get("HistoryID"), Gamename);
                 }
                 orderTransactions();
             }
@@ -225,28 +216,27 @@ public class DatabaseCon {
     }
 
     public void addTransaction(Number sender, Number recipient, double amount, Timestamp Time, String notiz, Number HistoryID, String Gamename) {
-        int t;
-        for(t=0; t < transactions.size(); t++){
-            if(transactions.get(t).TransactioID.equals(HistoryID)){
+        for (int t = 0; t < transactions.size(); t++) {
+            if (transactions.get(t).TransactioID.equals(HistoryID))
                 return;
-            }
         }
+
         String Empf = "DELETED";
-        String senders = "DELETED";
-        int e;
-        for(e=0; e < recipients.size(); e++){
-            if(recipients.get(e).KontoID.equals(recipient)){
+        for (int e = 0; e < recipients.size(); e++) {
+            if (recipients.get(e).KontoID.equals(recipient)) {
                 Empf = recipients.get(e).Name;
                 break;
             }
         }
-        int n;
-        for(n=0; n < recipients.size(); n++){
-            if(recipients.get(n).KontoID.equals(sender)){
+
+        String senders = "DELETED";
+        for (int n = 0; n < recipients.size(); n++) {
+            if (recipients.get(n).KontoID.equals(sender)) {
                 senders = recipients.get(n).Name;
                 break;
             }
         }
+
         Transaction newTran = new Transaction(senders, Empf, amount, Time, notiz, HistoryID, Gamename);
         transactions.add(newTran);
     }
@@ -254,16 +244,13 @@ public class DatabaseCon {
     public void connectRecipients(AccountSelectionActivity activity) {
         ColAccount.get().addOnCompleteListener(task -> {
             if (task.isSuccessful()) {
-                System.out.println("Start Recipients");
                 for (QueryDocumentSnapshot document : task.getResult()) {
                     Map<String, Object> m = document.getData();
-                    for(int i = 0; i < games.size(); i++){
-                        if(games.get(i).name.contentEquals(m.get("Game").toString())){
-                            addRecipient(m.get("Kontoname").toString(), m.get("Game").toString(), (Number)m.get("KontoID"));
-                        }
+                    for (int i = 0; i < games.size(); i++) {
+                        if (games.get(i).name.contentEquals(m.get("Game").toString()))
+                            addRecipient(m.get("Kontoname").toString(), m.get("Game").toString(), (Number) m.get("KontoID"));
                     }
                 }
-                System.out.println("Finish Recipients");
                 connectAccounts(activity);
             }
         });
@@ -279,9 +266,8 @@ public class DatabaseCon {
             if (task.isSuccessful()) {
                 for (QueryDocumentSnapshot document : task.getResult()) {
                     Map<String, Object> m = document.getData();
-                    addGame(m.get("Name").toString(), (Number)m.get("Admin"), Double.parseDouble(m.get("DefaultBalance").toString()));
+                    addGame(m.get("Name").toString(), (Number) m.get("Admin"), Double.parseDouble(m.get("DefaultBalance").toString()));
                 }
-                System.out.println("Games Connnect Finish");
                 connectRecipients(activity);
             }
         });
@@ -306,9 +292,9 @@ public class DatabaseCon {
 
     public void transferMoneyOnComplete(double Betrag, Number KontoIDempfaenger, Number senderKontoID, boolean FromAdmin) {
         ColAccount.document(KontoIDempfaenger.toString()).update("Geld", FieldValue.increment(Betrag));
-        if(!FromAdmin){
+
+        if (!FromAdmin)
             ColAccount.document(senderKontoID.toString()).update("Geld", FieldValue.increment(-1 * Betrag));
-        }
     }
 
     public String getName() {
@@ -316,51 +302,30 @@ public class DatabaseCon {
     }
 
     public Number getAdminName(String gameName) {
-
-        for(int i = 0; i < games.size(); i++){
-            if(games.get(i).name.equals(gameName)){
+        for (int i = 0; i < games.size(); i++) {
+            if (games.get(i).name.equals(gameName))
                 return games.get(i).adminID;
-            }
         }
+
         return -1;
     }
 
-    public String getNutzerName() {
-        return user.getNutzerName();
-    }
-
-    public String getNutzerPw() {
-        return user.getNutzerPW();
-    }
-
-    public Number getNutzerID() {
-        return user.getNutzerID();
-    }
-
-    public boolean getSucess() {
-        return success;
-    }
-
-    public void setSucess(boolean b) {
-        success = b;
-    }
-
     public Account getAccount(Integer i) {
-        if (i > accounts.size() - 1) {
+        if (i > accounts.size() - 1)
             return null;
-        }
+
         return accounts.get(i);
     }
-    public ArrayList<Account> getAllAccounts()
-    {
+
+    public ArrayList<Account> getAllAccounts() {
         return accounts;
     }
 
     @Nullable
     public Transaction getTransaction(Integer i) {
-        if (i > transactions.size() - 1) {
+        if (i > transactions.size() - 1)
             return null;
-        }
+
         return transactions.get(i);
     }
 
@@ -369,9 +334,9 @@ public class DatabaseCon {
         for (i = 0; this.getAccount(i) != null; i++) {
             accounts.get(i).AccountHistory.clear();
             for (j = 0; this.getTransaction(j) != null; j++) {
-                if ((transactions.get(j).sender.contentEquals(accounts.get(i).name) || transactions.get(j).recipient.contentEquals(accounts.get(i).name)) && transactions.get(j).Game.equals(accounts.get(i).gameName)) {
+                boolean isSenderOrRecipient = transactions.get(j).sender.contentEquals(accounts.get(i).name) || transactions.get(j).recipient.contentEquals(accounts.get(i).name);
+                if ((isSenderOrRecipient) && transactions.get(j).Game.equals(accounts.get(i).gameName))
                     accounts.get(i).AccountHistory.add(transactions.get(j));
-                }
             }
         }
     }
@@ -383,21 +348,18 @@ public class DatabaseCon {
         return games.get(i);
     }
 
-    public NutzerClass getUser() {
-        return user;
-    }
-
-    public Recipients getEmpfaenger(Integer i){
-        if (i > recipients.size() - 1) {
+    public Recipients getEmpfaenger(Integer i) {
+        if (i > recipients.size() - 1)
             return null;
-        }
+
         return recipients.get(i);
     }
 
-    private String getFromDeviceStorage(String variableName, String defaultValue) {
+    private String getFromDeviceStorage(String variableName) {
         String generated;
-        String id = sharedPreferences.getString(variableName, defaultValue);
-        if (id.equals(defaultValue)) {
+        String id = sharedPreferences.getString(variableName, "0123456789");
+
+        if (id.equals("0123456789")) {
             generated = new StringGenerator().nextString();
             SharedPreferences.Editor editor = sharedPreferences.edit();
             editor.putString(variableName, generated);
@@ -408,42 +370,34 @@ public class DatabaseCon {
         return generated;
     }
 
-    public void deleteAccount(String gameName, Number accountId)
-    {
-        int i= 0;
-        for( i = 0; i < accounts.size(); i++){
-            if(accounts.get(i).AccountID.equals(accountId)){
+    public void deleteAccount(String gameName, Number accountId) {
+        for (int i = 0; i < accounts.size(); i++) {
+            if (accounts.get(i).AccountID.equals(accountId)) {
                 Map<String, Object> m = new HashMap<>();
                 m.put("Game", accounts.get(i).gameName);
                 m.put("Geld", accounts.get(i).balance);
                 m.put("KontoID", accounts.get(i).AccountID);
                 m.put("Kontoname", accounts.get(i).name);
                 m.put("Nutzer", user.getNutzerID());
+
                 database.collection("deletedKonten").document(accounts.get(i).AccountID.toString()).set(m);
                 accounts.remove(i);
                 break;
             }
         }
+
         ColAccount.document(accountId.toString()).delete();
-        ColAccount.whereEqualTo("Game", gameName).get().addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
-
-            @Override
-            public void onComplete(@NonNull Task<QuerySnapshot> task) {
-                if (task.isSuccessful()) {
-                    if(task.getResult().isEmpty()){
-                       database.collection("Game").whereEqualTo("Name", gameName).get().addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
-
-                           @Override
-                           public void onComplete(@NonNull Task<QuerySnapshot> task) {
-                               if (task.isSuccessful()) {
-                                   for (QueryDocumentSnapshot document : task.getResult()) {
-                                       Number d = (Number)document.get("GameID");
-                                       database.collection("Game").document(d.toString()).delete();
-                                   }
-                               }
-                           }
-                       });
-                    }
+        ColAccount.whereEqualTo("Game", gameName).get().addOnCompleteListener(task -> {
+            if (task.isSuccessful()) {
+                if (task.getResult().isEmpty()) {
+                    database.collection("Game").whereEqualTo("Name", gameName).get().addOnCompleteListener(task1 -> {
+                        if (task1.isSuccessful()) {
+                            for (QueryDocumentSnapshot document : task1.getResult()) {
+                                Number d = (Number) document.get("GameID");
+                                database.collection("Game").document(d.toString()).delete();
+                            }
+                        }
+                    });
                 }
             }
         });
@@ -474,6 +428,7 @@ public class DatabaseCon {
         public String nextString() {
             for (int idx = 0; idx < buf.length; ++idx)
                 buf[idx] = symbols[random.nextInt(symbols.length)];
+
             return new String(buf);
         }
     }
